+++ packages/debug/service.go
@@ -2,24 +2,16 @@ package debug
 
 import (
 	"context"
-	"errors"
 	"fmt"
+	"time"
+
 	v1 "github.com/cloudogu/k8s-debug-mode-cr-lib/api/v1"
 	"github.com/sirupsen/logrus"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"time"
 
 	pbMaintenance "github.com/cloudogu/ces-control-api/generated/maintenance"
 	"github.com/cloudogu/ces-control-api/generated/types"
-)
-
-const (
-	maintenanceTitle          = "Service unavailable"
-	activateMaintenanceText   = "Activating debug mode"
-	deactivateMaintenanceText = "Deactivating debug mode"
-	logLevelDebug             = "DEBUG"
-	interErrMsg               = "internal error"
 )
 
 type defaultDebugModeService struct {
@@ -80,36 +72,6 @@ func (s *defaultDebugModeService) Enable(ctx context.Context, req *pbMaintenance
 	return &types.BasicResponse{}, nil
 }
 
-func (s *defaultDebugModeService) rollbackRestoreStartDisable(ctx context.Context, err error) error {
-	startErr := s.doguInterActor.StartAllDogus(ctx)
-	if startErr != nil {
-		err = errors.Join(wrapRollBackErr(startErr), err)
-	}
-
-	return s.rollbackRestoreDisable(ctx, err)
-}
-
-func (s *defaultDebugModeService) rollbackRestoreDisable(ctx context.Context, err error) error {
-	restoreErr := s.debugModeRegistry.RestoreDoguLogLevels(ctx)
-	if restoreErr != nil {
-		err = errors.Join(wrapRollBackErr(restoreErr), err)
-	}
-
-	return s.rollbackDisable(ctx, err)
-}
-
-func (s *defaultDebugModeService) rollbackDisable(ctx context.Context, err error) error {
-	rollbackErr := s.debugModeRegistry.Disable(ctx)
-	if rollbackErr != nil {
-		err = errors.Join(wrapRollBackErr(rollbackErr), err)
-	}
-	return err
-}
-
-func wrapRollBackErr(err error) error {
-	return fmt.Errorf("rollback error: %w", err)
-}
-
 // Disable returns an error because the method is unimplemented.
 func (s *defaultDebugModeService) Disable(ctx context.Context, _ *pbMaintenance.ToggleDebugModeRequest) (*types.BasicResponse, error) {
 	logrus.Info("Starting to disable debug-mode...")
@@ -137,7 +99,7 @@ func (s *defaultDebugModeService) Status(ctx context.Context, _ *types.BasicRequ
 		return nil, fmt.Errorf("ERROR: failed to get debug-mode: %q", err)
 	}
 
-	return &pbMaintenance.DebugModeStatusResponse{IsEnabled: debugMode.Status.Phase != v1.DebugModeStatusCompleted, DisableAtTimestamp: debugMode.Spec.DeactivateTimestamp.Unix()}, nil
+	return &pbMaintenance.DebugModeStatusResponse{IsEnabled: debugMode.Status.Phase != v1.DebugModeStatusCompleted, DisableAtTimestamp: debugMode.Spec.DeactivateTimestamp.UnixMilli()}, nil
 }
 
 func noInheritCancel(_ context.Context) (context.Context, context.CancelFunc) {