+++ CHANGELOG.md
@@ -10,6 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#76] get & set backup-schedule
 - [#76] get retention-policy
 
+## [v1.7.2] - 2025-09-17
+### Fixed
+- [#71] remove unused maintenance mode code.
+
+## [v1.7.1] - 2025-08-29
+### Fixed
+- [#70] use UnixMilli instead of Unix Time for DebugMode Status request
+
 ## [v1.7.0] - 2025-08-14
 ### Added
 - [#66] Create debugMode-CRs